Fix bug in engine. New au::str method for double values. Minor refactoring
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3365 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/au/Cronometer.cpp
@@ -48,11 +48,11 @@ double Cronometer::diffTime()
 {
     struct timeval t2;
     
-    if ( gettimeofday(&t2,NULL) != 0)
+    if ( gettimeofday(&t2,NULL) != 0 )
         LM_X(1, ("gettimeofday failed"));
     
-    
-    double diff = ( (double) t2.tv_sec ) + ((double)t2.tv_usec / 1000000.0) - ( (double) t.tv_sec ) - ((double)t.tv_usec / 1000000.0);
+    double diff = ( (double) t2.tv_sec ) + ((double)t2.tv_usec / (double)1000000.0 ) 
+    - ( (double) t.tv_sec ) - ((double)t.tv_usec / (double)1000000.0);
 
 /*    
     {
